pyproject.toml
[tool.poetry]
name = "resume"
version = "0.1.0"
description = ""
authors = ["Kamil Urbanek <urbanek.kamil@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
invoke = "^1.7.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
